fix(project-intelligence): dedupe sampler CIM spawn plumbing to clear SonarCloud duplication gate (refs #1155)
PR #1166's Sonar quality gate failed on new_duplicated_lines_density
(22.6% > 3%): adding an identical unrefChildAndPipes(child) line to
both near-identical spawn+collect-stdout blocks in
clients/resource-sampler.ts pushed duplication over the threshold —
the same duplication shape #1153/#1160 already solved for the orphan
reaper via spawnCollectStdout.

Promote spawnCollectStdout (spawn -> unref -> pipe stdout -> resolve
on close/error, resolving to "" on any failure) from
clients/instance-reaper.ts into the shared clients/child-unref.ts
alongside unrefChildAndPipes, and use it at both resource-sampler.ts
spawn sites. Both call sites become parse-only around the shared
helper; instance-reaper.ts now imports spawnCollectStdout instead of
defining it locally, with zero behavior change there.
